Adds support for regex flags
This change adds support for regex flags. Currently if we want to supply flags such as i for ignore case, we must either pass in the RegExp object /\w+/i or pass it in as a flag parameter. Since we return the regex as a string, we cannot pass the Regex object via getInput. Instead this change proposes passing in flags as a separate input.